ui: move editable combo control to Controls namespace
