+++ Views/Controls/EditableComboWithAdd.xaml
@@ -1,4 +1,4 @@
-<UserControl x:Class="Facturon.App.Views.EditableComboWithAdd"
+<UserControl x:Class="Facturon.App.Views.Controls.EditableComboWithAdd"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
     <UserControl.Resources>
