+++ Views/InvoiceDetailView.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:controls="clr-namespace:Facturon.App.Views"
+             xmlns:controls="clr-namespace:Facturon.App.Views.Controls"
              mc:Ignorable="d">
     <UserControl.Resources>
         <DataTemplate x:Key="SupplierInlineCreate">
